Move convergence class definition into its own file
